refactor(tools): move BashTool into cube-tools/cube-bash-tool optional package
Follows the same pattern as cube-browser-tool and cube-computer-tool: concrete
tool implementations live in standalone optional packages under cube-tools/,
not in the cube-standard src tree.

- Adds cube-tools/cube-bash-tool/ with pyproject.toml, Makefile, and uv.lock
- Moves src/cube/tools/bash.py → cube-tools/cube-bash-tool/src/cube_bash_tool/
- Moves tests/test_bash_tool.py → cube-tools/cube-bash-tool/tests/
- Reverts src/cube/tools/__init__.py to empty stub
- Updates cube-tools/README.md package table

Tool.NON_INTERACTIVE_ENV and Tool._truncate_output stay on the base class
(cube-standard) as shared helpers for any container-backed tool.
