+++ src/cube/tools/__init__.py
@@ -1,5 +1,1 @@
 """Tool domain subclasses for cube-standard benchmarks."""
-
-from cube.tools.bash import BashTool, BashToolConfig
-
-__all__ = ["BashTool", "BashToolConfig"]
